Separated number field and time field as a subclass
